LayoutTests/websocket/tests/url-with-empty-query-expected.txt

Make sure handshake with URL with empty query components success.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

url=ws://127.0.0.1:8880/websocket/tests/echo-location?
WebSocket is open
received:ws://127.0.0.1:8880/websocket/tests/echo-location?
WebSocket is closed
PASS handshake_success is true
PASS ws_location is url
PASS successfullyParsed is true

TEST COMPLETE
